requirements.txt
# Backend Dependencies
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
sqlalchemy>=2.0.23
alembic>=1.12.1
pydantic>=2.5.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
requests>=2.31.0

# Frontend Dependencies (תואמות Python 3.12)
PySide6>=6.6.0
PySide6-Addons>=6.6.0
PySide6-Essentials>=6.6.0

# AI & ML Dependencies (גירסאות חדשות יותר)
openai>=1.3.7
langchain>=0.1.0
chromadb>=0.4.17
sentence-transformers>=2.2.2
ollama>=0.1.7

# Database & Storage (sqlite3 מובנה ב-Python)
cloudinary>=1.36.0

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
pylint==3.0.3
python-dotenv==1.0.0

# Utilities
python-dateutil==2.8.2
Pillow==10.1.0
jinja2==3.1.2